api-python/requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
six==1.16.0
Werkzeug==2.1.2
